project_root/Database/requirements.txt
fastapi==0.115.5
jose==1.0.0
passlib==1.7.4
pydantic==1.10.2
pymongo==4.10.1
python-dotenv==1.0.1
